Parent Interaction class added. graci-to-bitsi SOC interface added
