+++ graci/core/libs.py
@@ -105,6 +105,9 @@ bitsi_registry = {
     'bitsi_finalise'          : [],
     'transition_density_mrci' : ['int32','int32','int32','int32','int32',
                                  'int32','double','string','string',
+                                 'string','string'],
+    'soc_mrci'                : ['int32','int32','int32','int32','int32',
+                                 'int32','double','string','string',
                                  'string','string']
 }
 
@@ -112,6 +115,8 @@ bitsi_intent   = {
     'bitsi_initialise'        : ['in','in','in','in','in','in','in'],
     'bitsi_finalise'          : [],
     'transition_density_mrci' : ['in','in','in','in','in','in','out',
+                                 'in','in','in','in'],
+    'soc_mrci'                : ['in','in','in','in','in','in','out',
                                  'in','in','in','in']
 }
