+++ graci/io/parse.py
@@ -264,7 +264,7 @@ def check_input(run_list):
                     
         # init/final_states and i/fstate_array need to be lists, also:
         # internal state ordering is 0->n-1, vs. 1->n for input
-        if type(obj).__name__ == 'Transition':
+        if type(obj).__name__ == 'Transition' or type(obj).__name__ == 'Spinorbit':
             if obj.init_states is not None:
                 if not isinstance(obj.init_states, (list, np.ndarray)):
                     obj.init_states = [obj.init_states]
